refactor(#47): excutor business layer 분리
