+++ api/src/main/java/co/kr/cocomu/executor/producer/CodeExecutionProducer.java
@@ -1,4 +1,4 @@
-package co.kr.cocomu.executor.service;
+package co.kr.cocomu.executor.producer;
 
 
 import co.kr.cocomu.executor.dto.message.CodeExecutionMessage;
