+++ api/src/test/java/co/kr/cocomu/executor/producer/CodeExecutionProducerTest.java
@@ -1,4 +1,4 @@
-package co.kr.cocomu.executor.service;
+package co.kr.cocomu.executor.producer;
 
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
